Chart Templates and functions splitted
