+++ src/App.js
@@ -1,10 +1,7 @@
 import React from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import Home from './Components/Home'
-import MonthlyReport from './Components/Templates/PieChartT';
-import Bouncerate from './Components/Bouncerate';
-import WorldMap from './Components/WorldMap';
-import FormSubmit from './Components/FormSubmit';
+
 
 function App() {
   return (
@@ -12,10 +9,7 @@ function App() {
       <BrowserRouter>
         <Routes>
           <Route path="/" element={<Home />} />
-          <Route path="/monthly-report" element={<MonthlyReport />} />
-          <Route path="/bounce-rate" element={<Bouncerate/>}/>
-          <Route path="/map" element={<WorldMap/>}/>
-          <Route path="/form-report" element={<FormSubmit />} />
+
         </Routes>
       </BrowserRouter>
     </div>
