+++ src/Components/Templates/PieChartT.js
@@ -2,7 +2,7 @@ import {useState,useEffect} from 'react';
 import { PieChart } from "@mui/x-charts/PieChart";
 
 
-const MonthlyReport = ({data}) => {
+const PieChartT = ({data}) => {
   useEffect(()=>{
     console.log({data});
   })
@@ -19,4 +19,4 @@ const MonthlyReport = ({data}) => {
   );
 }
 
-export default MonthlyReport;
+export default PieChartT;
